Fix lint warnings around hooks imports
