+++ .changelog/2417.trivial.md
@@ -1,0 +1,1 @@
+Fix lint warnings around hooks imports
