+++ src/app/pages/ConsensusBlockDetailPage/ConsensusBlockEventsCard.tsx
@@ -3,7 +3,7 @@ import { useGetConsensusEvents } from '../../../oasis-nexus/api'
 import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE as limit } from '../../../config'
 import { LinkableCardLayout } from '../../components/LinkableCardLayout'
 import { useSearchParamsPagination } from '../../components/Table/useSearchParamsPagination'
-import { ConsensusBlockDetailsContext } from '.'
+import { ConsensusBlockDetailsContext } from './types'
 import { ConsensusEventsList } from '../../components/ConsensusEvents/ConsensusEventsList'
 import { eventsContainerId } from '../../utils/tabAnchors'
 import { getConsensusEventTypeFilteringParam, useConsensusEventTypeParam } from '../../hooks/useCommonParams'
