+++ src/app/pages/ConsensusBlockDetailPage/hooks.tsx
@@ -1,0 +1,4 @@
+import { useOutletContext } from 'react-router-dom'
+import { ConsensusBlockDetailsContext } from './types'
+
+export const useConsensusBlockDetailsProps = () => useOutletContext<ConsensusBlockDetailsContext>()
