+++ src/app/pages/ConsensusBlockDetailPage/types.ts
@@ -1,0 +1,6 @@
+import { ConsensusScope } from '../../../types/searchScope'
+
+export type ConsensusBlockDetailsContext = {
+  scope: ConsensusScope
+  blockHeight?: number
+}
