+++ src/app/pages/NFTInstanceDashboardPage/InstanceDetailsCard.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 import { Card, CardContent } from '@oasisprotocol/ui-library/src/components/cards'
-import { useAccount } from '../RuntimeAccountDetailsPage/hook'
+import { useAccount } from '../RuntimeAccountDetailsPage/hooks'
 import { TextSkeleton } from '../../components/Skeleton'
 import { StyledDescriptionList } from '../../components/StyledDescriptionList'
 import { useTranslation } from 'react-i18next'
