+++ src/app/pages/NFTInstanceDashboardPage/hooks.tsx
@@ -1,0 +1,4 @@
+import { useOutletContext } from 'react-router-dom'
+import { NftDashboardContext } from './types'
+
+export const useNftDetailsProps = () => useOutletContext<NftDashboardContext>()
