+++ src/app/pages/NFTInstanceDashboardPage/types.ts
@@ -1,0 +1,7 @@
+import { RuntimeScope } from '../../../types/searchScope'
+
+export type NftDashboardContext = {
+  scope: RuntimeScope
+  address: string
+  instanceId: string
+}
