+++ src/app/pages/RuntimeAccountDetailsPage/AccountEventsCard.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react'
 import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE as limit } from '../../../config'
 import { LinkableCardLayout } from '../../components/LinkableCardLayout'
 import { RuntimeEventsDetailedList } from '../../components/RuntimeEvents/RuntimeEventsDetailedList'
-import { useAccountEvents } from './hook'
-import { RuntimeAccountDetailsContext } from '.'
+import { useAccountEvents } from './hooks'
+import { RuntimeAccountDetailsContext } from './types'
 import { eventsContainerId } from '../../utils/tabAnchors'
 import { RuntimeEventTypeFilter } from '../../components/RuntimeEvents/RuntimeEventTypeFilter'
 import { CardDivider } from '../../components/Divider'
