+++ src/app/pages/RuntimeAccountDetailsPage/AccountNFTCollectionCard.tsx
@@ -7,14 +7,14 @@ import { Typography } from '@oasisprotocol/ui-library/src/components/typography'
 import { Skeleton } from '@oasisprotocol/ui-library/src/components/ui/skeleton'
 import { ErrorBoundary } from '../../components/ErrorBoundary'
 import { LinkableDiv } from '../../components/PageLayout/LinkableDiv'
-import { RuntimeAccountDetailsContext } from './index'
+import { RuntimeAccountDetailsContext } from './types'
 import { AccountLink } from 'app/components/Account/AccountLink'
 import { CopyToClipboard } from 'app/components/CopyToClipboard'
 import { AddressLoaderData, RouteUtils } from 'app/utils/route-utils'
 import { ImageListItemImage } from '../TokenDashboardPage/ImageListItemImage'
 import { CardEmptyState } from '../../components/CardEmptyState'
 import { TablePagination } from '../../components/Table/TablePagination'
-import { useAccountTokenInventory } from '../TokenDashboardPage/hook'
+import { useAccountTokenInventory } from '../TokenDashboardPage/hooks'
 import { EvmNft } from 'oasis-nexus/api'
 import { SearchScope } from '../../../types/searchScope'
 import { NFTCollectionLink, NFTInstanceLink } from '../TokenDashboardPage/NFTLinks'
