+++ src/app/pages/RuntimeAccountDetailsPage/AccountTransactionsCard.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react'
 import { LinkableCardLayout } from '../../components/LinkableCardLayout'
 import { RuntimeTransactions } from '../../components/Transactions'
 import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE } from '../../../config'
-import { useAccountTransactions } from './hook'
-import { RuntimeAccountDetailsContext } from './index'
+import { useAccountTransactions } from './hooks'
+import { RuntimeAccountDetailsContext } from './types'
 import { useScreenSize } from '../../hooks/useScreensize'
 import { RuntimeTransactionMethodFilter } from '../../components/Transactions/RuntimeTransactionMethodFilter'
 import { transactionsContainerId } from '../../utils/tabAnchors'
