+++ src/app/pages/RuntimeAccountDetailsPage/ContractCodeCard.tsx
@@ -2,14 +2,14 @@ import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Card, CardContent } from '@oasisprotocol/ui-library/src/components/cards'
 import { LinkableDiv } from '../../components/PageLayout/LinkableDiv'
-import { useAccount } from './hook'
+import { useAccount } from './hooks'
 import { CardEmptyState } from '../../components/CardEmptyState'
-import { TokenDashboardContext } from '../TokenDashboardPage'
+import { TokenDashboardContext } from '../TokenDashboardPage/types'
 import { CodeDisplay } from '../../components/CodeDisplay'
 import { codeContainerId } from '../../utils/tabAnchors'
 import { Typography } from '@oasisprotocol/ui-library/src/components/typography'
 import { VerificationIcon } from 'app/components/ContractVerificationIcon'
-import { useTokenInfo } from '../TokenDashboardPage/hook'
+import { useTokenInfo } from '../TokenDashboardPage/hooks'
 import { base64ToHex } from '../../utils/helpers'
 import { Separator } from '@oasisprotocol/ui-library/src/components/ui/separator'
