+++ src/app/pages/RuntimeBlockDetailPage/RuntimeBlockEventsCard.tsx
@@ -4,7 +4,7 @@ import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE as limit } from '../../../config'
 import { LinkableCardLayout } from '../../components/LinkableCardLayout'
 import { useSearchParamsPagination } from '../../components/Table/useSearchParamsPagination'
 import { RuntimeEventsDetailedList } from '../../components/RuntimeEvents/RuntimeEventsDetailedList'
-import { RuntimeBlockDetailsContext } from '.'
+import { RuntimeBlockDetailsContext } from './types'
 import { eventsContainerId } from '../../utils/tabAnchors'
 import { getRuntimeEventTypeFilteringParam } from '../../hooks/useCommonParams'
 import { RuntimeEventTypeFilter } from '../../components/RuntimeEvents/RuntimeEventTypeFilter'
