+++ src/app/pages/RuntimeBlockDetailPage/RuntimeBlockTransactionsCard.tsx
@@ -5,7 +5,7 @@ import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE } from '../../../config'
 import { useGetRuntimeTransactions } from '../../../oasis-nexus/api'
 import { RuntimeTransactions } from '../../components/Transactions'
 import { AppErrors } from '../../../types/errors'
-import { RuntimeBlockDetailsContext } from '.'
+import { RuntimeBlockDetailsContext } from './types'
 import { useScreenSize } from '../../hooks/useScreensize'
 import { RuntimeTransactionMethodFilter } from '../../components/Transactions/RuntimeTransactionMethodFilter'
 import { getRuntimeTransactionMethodFilteringParam } from '../../components/RuntimeTransactionMethod'
