+++ src/app/pages/RuntimeBlockDetailPage/hooks.tsx
@@ -1,0 +1,4 @@
+import { useOutletContext } from 'react-router-dom'
+import { RuntimeBlockDetailsContext } from './types'
+
+export const useRuntimeBlockDetailsProps = () => useOutletContext<RuntimeBlockDetailsContext>()
