+++ src/app/pages/RuntimeTransactionDetailPage/index.tsx
@@ -33,7 +33,7 @@ import { SimpleJsonCode } from '../../components/CodeDisplay/SimpleJsonCode'
 import { isRoflTransaction } from '../../utils/transaction'
 import { RoundedBalance } from 'app/components/RoundedBalance'
 import * as oasis from '@oasisprotocol/client'
-import { useTokenTransfers } from '../TokenDashboardPage/hook'
+import { useTokenTransfers } from '../TokenDashboardPage/hooks'
 import { TokenTypeTag } from 'app/components/Tokens/TokenList'
 import { LinkableDiv } from 'app/components/PageLayout/LinkableDiv'
 import { EventBalance } from 'app/components/Tokens/TokenTransfers'
