+++ src/app/pages/TokenDashboardPage/TokenDetailsCard.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react'
 import { Link as RouterLink } from 'react-router-dom'
 import { Card, CardContent } from '@oasisprotocol/ui-library/src/components/cards'
 import { Link } from '@oasisprotocol/ui-library/src/components/link'
-import { useTokenInfo } from './hook'
-import { useAccount } from '../RuntimeAccountDetailsPage/hook'
+import { useTokenInfo } from './hooks'
+import { useAccount } from '../RuntimeAccountDetailsPage/hooks'
 import { TextSkeleton } from '../../components/Skeleton'
 import { StyledDescriptionList } from '../../components/StyledDescriptionList'
 import { useScreenSize } from '../../hooks/useScreensize'
