+++ src/app/pages/TokenDashboardPage/TokenInventoryCard.tsx
@@ -2,10 +2,10 @@ import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { LinkableCardLayout } from '../../components/LinkableCardLayout'
 import { CardEmptyState } from '../../components/CardEmptyState'
-import { TokenDashboardContext } from './index'
+import { TokenDashboardContext } from './types'
 import { RouteUtils } from '../../utils/route-utils'
 import { TablePagination } from '../../components/Table/TablePagination'
-import { useTokenInventory } from './hook'
+import { useTokenInventory } from './hooks'
 import { ImageListItemImage } from './ImageListItemImage'
 import { NFTInstanceLink, NFTOwnerLink } from './NFTLinks'
 import { EvmNft } from 'oasis-nexus/api'
