+++ src/app/pages/TokenDashboardPage/TokenSupplyCard.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { SnapshotTextCard } from '../../components/Snapshots/SnapshotCard'
-import { useTokenInfo } from './hook'
+import { useTokenInfo } from './hooks'
 import { Skeleton } from '@oasisprotocol/ui-library/src/components/ui/skeleton'
 import { RuntimeScope } from '../../../types/searchScope'
 import { RoundedBalance } from '../../components/RoundedBalance'
