+++ src/app/pages/TokenDashboardPage/TokenTypeCard.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Typography } from '@oasisprotocol/ui-library/src/components/typography'
 import { SnapshotCard } from '../../components/Snapshots/SnapshotCard'
-import { useTokenInfo } from './hook'
+import { useTokenInfo } from './hooks'
 import { Skeleton } from '@oasisprotocol/ui-library/src/components/ui/skeleton'
 import { getTokenTypeDescription, getTokenTypeStrictName } from '../../../types/tokens'
 import { RuntimeScope } from '../../../types/searchScope'
