+++ src/app/pages/TokenDashboardPage/types.ts
@@ -1,0 +1,6 @@
+import { RuntimeScope } from '../../../types/searchScope'
+
+export type TokenDashboardContext = {
+  scope: RuntimeScope
+  address: string
+}
