+++ src/routes.tsx
@@ -4,13 +4,12 @@ import { RuntimeBlocksPage } from './app/pages/RuntimeBlocksPage'
 import { RuntimeTransactionsPage } from './app/pages/RuntimeTransactionsPage'
 import { RuntimeTransactionDetailPage } from './app/pages/RuntimeTransactionDetailPage'
 import { ParatimeDashboardPage } from './app/pages/ParatimeDashboardPage'
-import { RuntimeBlockDetailPage, useRuntimeBlockDetailsProps } from './app/pages/RuntimeBlockDetailPage'
+import { RuntimeBlockDetailPage } from './app/pages/RuntimeBlockDetailPage'
+import { useRuntimeBlockDetailsProps } from './app/pages/RuntimeBlockDetailPage/hooks'
 import { RuntimeBlockTransactionsCard } from './app/pages/RuntimeBlockDetailPage/RuntimeBlockTransactionsCard'
 import { RuntimeBlockEventsCard } from './app/pages/RuntimeBlockDetailPage/RuntimeBlockEventsCard'
-import {
-  RuntimeAccountDetailsPage,
-  useRuntimeAccountDetailsProps,
-} from './app/pages/RuntimeAccountDetailsPage'
+import { RuntimeAccountDetailsPage } from './app/pages/RuntimeAccountDetailsPage'
+import { useRuntimeAccountDetailsProps } from './app/pages/RuntimeAccountDetailsPage/hooks'
 import { AccountTransactionsCard } from './app/pages/RuntimeAccountDetailsPage/AccountTransactionsCard'
 import { AccountTokensCard } from './app/pages/RuntimeAccountDetailsPage/AccountTokensCard'
 import { SearchResultsPage } from './app/pages/SearchResultsPage'
@@ -34,13 +33,15 @@ import { useRequiredScopeParam } from './app/hooks/useScopeParam'
 import { TokensPage } from './app/pages/TokensOverviewPage'
 import { AccountEventsCard } from 'app/pages/RuntimeAccountDetailsPage/AccountEventsCard'
 import { ContractCodeCard } from './app/pages/RuntimeAccountDetailsPage/ContractCodeCard'
-import { TokenDashboardPage, useTokenDashboardProps } from './app/pages/TokenDashboardPage'
+import { TokenDashboardPage } from './app/pages/TokenDashboardPage'
+import { useTokenDashboardProps } from './app/pages/TokenDashboardPage/hooks'
 import { AccountTokenTransfersCard } from './app/pages/RuntimeAccountDetailsPage/AccountTokenTransfersCard'
 import { AccountNFTCollectionCard } from './app/pages/RuntimeAccountDetailsPage/AccountNFTCollectionCard'
 import { TokenTransfersCard } from './app/pages/TokenDashboardPage/TokenTransfersCard'
 import { TokenHoldersCard } from './app/pages/TokenDashboardPage/TokenHoldersCard'
 import { TokenInventoryCard } from './app/pages/TokenDashboardPage/TokenInventoryCard'
-import { NFTInstanceDashboardPage, useNftDetailsProps } from './app/pages/NFTInstanceDashboardPage'
+import { NFTInstanceDashboardPage } from './app/pages/NFTInstanceDashboardPage'
+import { useNftDetailsProps } from './app/pages/NFTInstanceDashboardPage/hooks'
 import { NFTMetadataCard } from './app/pages/NFTInstanceDashboardPage/NFTMetadataCard'
 import { NFTTokenTransfersCard } from './app/pages/NFTInstanceDashboardPage/NFTTokenTransfersCard'
 import { ConsensusDashboardPage } from 'app/pages/ConsensusDashboardPage'
@@ -53,7 +54,8 @@ import { DelegatorsCard } from './app/pages/ValidatorDetailsPage/DelegatorsCard'
 import { SearchScope } from './types/searchScope'
 import { ProposalDetailsPage } from './app/pages/ProposalDetailsPage'
 import { ConsensusBlocksPage } from './app/pages/ConsensusBlocksPage'
-import { ConsensusBlockDetailPage, useConsensusBlockDetailsProps } from './app/pages/ConsensusBlockDetailPage'
+import { ConsensusBlockDetailPage } from './app/pages/ConsensusBlockDetailPage'
+import { useConsensusBlockDetailsProps } from './app/pages/ConsensusBlockDetailPage/hooks'
 import { ConsensusBlockTransactionsCard } from './app/pages/ConsensusBlockDetailPage/ConsensusBlockTransactionsCard'
 import { ConsensusBlockEventsCard } from './app/pages/ConsensusBlockDetailPage/ConsensusBlockEventsCard'
 import { ConsensusAccountsPage } from './app/pages/ConsensusAccountsPage'
